[wallet-ext] Data flow UX improvements (#11294)
## Description 

This makes a few changes to the wallet:
- Reworks some of the feature gating logic to just expose growthbook
instances instead of wrapped instances. I think we could do a lot more
here to improve growthbook, I think having two instances is probably
actually fine, and maybe we should just poll from each of them to keep
features up-to-date and avoid sticky features.
- Re-enable refetch on mount, which broke things when persistent caching
was introduced.
- Add a feature flag that polls on the activity page.
- Add a feature flag that changes the request type for staking
transactions, and will just return when we have the effects cert, and
pass that through state to the receipt page. This improves the response
times pretty significantly when performing staking operations (shaves
~5s off of the flow, end-to-end).

## Test Plan 

Ran locally.