Make rocksdb sampler sample per column family (#11273)
## Description 

Today, when we create a `DBMap` we invoke this call:
```
DBMap {
            rocksdb: db.clone(),
            opts: opts.clone(),
            _phantom: PhantomData,
            cf: opt_cf.to_string(),
            db_metrics: db_metrics_cloned,
            _metrics_task_cancel_handle: Arc::new(sender),
            read_sample_interval: db.read_sampling_interval(),
            write_sample_interval: db.write_sampling_interval(),
            iter_bytes_sample_interval: db.iter_bytes_sampling_interval(),
            iter_latency_sample_interval: db.iter_latency_sampling_interval(),
        }
```
where in `read_sampling_interval()` we create a clone of the passed in
sampling interval for every `DBMap` of the table:
```
 pub fn read_sampling_interval(&self) -> SamplingInterval {
        match self {
            Self::DBWithThreadMode(d) => d.metric_conf.read_sample_interval.clone(),
            Self::OptimisticTransactionDB(d) => d.metric_conf.read_sample_interval.clone(),
        }
    }
```
This effectively means we have one sampling interval which samples
operations across all column families and all operations of the db. In
such a case, a hot cf (with a lot of operations) would not let us
capture perf data for other cfs. The idea in this PR is to create a
different sampling interval for each DBMap (cf) so we can capture perf
for all of them.
## Test Plan 

Existing tests