[fix] fixing failing randomized tests (#11331)
## Description 

This PR is fixing an issue on the behaviour of
`this_cert_parents_with_slow_nodes` which is used in randomized tests -
and in other consensus unit tests as well. The method is picking the
parents of a certificate supporting some failure modes (slow nodes). So
basically , it does pick a subset of the given parents when slow nodes
exist. However, it does not verifying that the parents honour the `2f +
1` requirement by stake. This is now fixed.

Now the randomized tests [should not
fail](https://github.com/MystenLabs/sui/actions/runs/4793242232) anymore

## Test Plan 

Existing randomized_tests should not fail.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes