Sui surfer (#11268)
This PR adds a sui-surfer tool that:
1. It starts a network locally
2. Publishes a package to the network. I wrote a package that contains a
bunch of object related building block entry functions. This is what it
publishes. But we could publish any number of packages to the network.
3. It starts 20 (configurable) independent account threads. Each thread
manages its gas and owned objects. All threads share an immutable/shared
object pool (each will contribute to it when new objects are generated).
3. Each thread crawl all entry functions in the network in a random
order, and for each function argument, it attempts to see if it could
match an object in the pool, and if so use them. For non-object
primitive arguments it just generates a random value for it.
4. Uses a strategy pattern, which means that one can plug in a different
pattern to decide what functions to call during surfing.