[ConsensusAdapter] monitor observed latency and submission interval separately (#11584)
## Description 

It seems currently observed latency is < 5s most of the time, so their
actual values are not exposed via metrics. Re-using the
`sequencing_estimated_latency` metric to expose the observed latency, so
we can tune the threshold for resubmission better.

## Test Plan 

CI.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes