[Part 1/2] Delete test utils messages (#11717)
This PR is the first of two (to avoid the PR growing too big) to remove
test-utils/messages eventually.
It makes the following changes:
1. Added a new sui-test-transaction-builder crate that allows us to
create all kinds of test transactions. This cannot live inside sui-types
because of package publishes, which requires compiling and verifying the
package (which needs to depends on sui-move-build). Migrated some of the
functions in messages.rs to use it.
2. Moved sign_transaction from TestCluster to WalletContext because it
belongs there.