[wallet-ext] remove the Ledger integration feature flag (#11914)
## Description 
Just doing some spring cleaning since we've released the Ledger
integration. No need for this feature flag anymore!

Note: The feature flag itself won't be deleted in Growthbook for a while
since this PR needs to make it to a wallet release and people might be
still be on old wallet versions

## Test Plan 
- Manual testing (Ledger feature is always on regardless of the flag
setting)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A