[sui cli] add validator command to print out raw gas price setting txn (#12034)
## Description 

This PR adds a validator command that prints out the base64
representation of a transaction that updates a validator's gas price
quote to a new value. I considered making this part of the original
`UpdateGasPrice` command but decided it's different enough to be its own
command, primarily because the active address of the wallet context can
be someone other than the transaction sender in the case of the new
command.
 
## Test Plan 

Added a test.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes