Remove days from countdown timeAgo  (#12141)
## Description 
Removing days from the countdown `timeAgo` helper. The wallet `Staking
Rewards Redeemable` was the only place this is visible, and this should
be in hours, not days.

<img width="405" alt="Screenshot 2023-05-22 at 12 19 39 PM"
src="https://github.com/MystenLabs/sui/assets/126525197/9d07f282-5bde-49aa-a01e-1667dcb84e69">


Describe the changes or additions included in this PR.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes