[ts-sdk] Update SDK to use new internals, add `subscribeTransaction` (#12008)
## Description 

This updates the internals of the ts-sdk to use `@open-rpc/client-js`.
This package supports different transports natively and is smaller than
our previous `jayson` / `rpc-websockets` implementation. As part of this
I also cleaned up a lot of our websocket code to make it easier to add
new methods, changed the API to return an unsubscribe function (which is
easier for most folks to reason about), and added the missing
`subscribeTransaction` method.

## Test Plan 

Added new tests for `subscribeTransaction`.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes