[Core] Add support for chain-specific configs (#12183)
## Description 

The goal is to be able to roll out features for certain chains and not
others.

To accomplish this, we introduce a `Chain` enum, with a getter function
to map the hard coded checkpoint identifier (i.e. the digest of the
genesis checkpoint) to the `Chain`, and have all callsites where the
protocol config is materialized to provide the chain identifier.

With the `Chain` being available in the
`ProtocolConfig::get_for_version`, we can then make use of this in the
protocol version map.

## Test Plan 

Tested against a Mainnet node to ensure that the chain is correctly
identified. Also locally enabled a feature for all chains but mainnet in
the highest protocol version, and saw that the feature was active.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes